@@ -1,4 +1,3 @@
-
 import 'dart:math' as math;
 import 'dart:ui';
 
@@ -7,16 +6,15 @@ import 'package:flutter/widgets.dart';
 import 'polygon_path_drawer.dart';
 
 class PolygonBorder extends ShapeBorder {
-
   const PolygonBorder({
-    @required this.sides,
+    required this.sides,
     this.rotate = 0.0,
     this.borderRadius = 0.0,
     this.border = BorderSide.none,
-  }) : assert(sides != null),
-       assert(rotate != null),
-       assert(borderRadius != null),
-       assert(border != null);
+  })  : assert(sides != null),
+        assert(rotate != null),
+        assert(borderRadius != null),
+        assert(border != null);
 
   final int sides;
   final double rotate;
@@ -33,18 +31,18 @@ class PolygonBorder extends ShapeBorder {
       borderRadiusAngle: borderRadius,
     );
 
-    return PolygonPathDrawer(size: Size.fromRadius(radius), specs: specs).draw()
+    return PolygonPathDrawer(size: Size.fromRadius(radius), specs: specs)
+        .draw()
         .shift(Offset(rect.center.dx - radius, rect.center.dy - radius));
   }
 
-
   @override
-  ShapeBorder lerpFrom(ShapeBorder a, double t) {
+  ShapeBorder? lerpFrom(ShapeBorder? a, double t) {
     if (a is PolygonBorder && a.sides == sides) {
       return PolygonBorder(
         sides: sides,
-        rotate: lerpDouble(a.rotate, rotate, t),
-        borderRadius: lerpDouble(a.borderRadius, borderRadius, t),
+        rotate: lerpDouble(a.rotate, rotate, t)!,
+        borderRadius: lerpDouble(a.borderRadius, borderRadius, t)!,
         border: BorderSide.lerp(a.border, border, t),
       );
     } else {
@@ -52,14 +50,13 @@ class PolygonBorder extends ShapeBorder {
     }
   }
 
-
   @override
-  ShapeBorder lerpTo(ShapeBorder b, double t) {
+  ShapeBorder? lerpTo(ShapeBorder? b, double t) {
     if (b is PolygonBorder && b.sides == sides) {
       return PolygonBorder(
         sides: sides,
-        rotate: lerpDouble(rotate, b.rotate, t),
-        borderRadius: lerpDouble(borderRadius, b.borderRadius, t),
+        rotate: lerpDouble(rotate, b.rotate, t)!,
+        borderRadius: lerpDouble(borderRadius, b.borderRadius, t)!,
         border: BorderSide.lerp(border, b.border, t),
       );
     } else {
@@ -68,7 +65,7 @@ class PolygonBorder extends ShapeBorder {
   }
 
   @override
-  void paint(Canvas canvas, Rect rect, {TextDirection textDirection}) {
+  void paint(Canvas canvas, Rect rect, {TextDirection? textDirection}) {
     switch (border.style) {
       case BorderStyle.none:
         break;
@@ -81,26 +78,31 @@ class PolygonBorder extends ShapeBorder {
   }
 
   @override
-  Path getInnerPath(Rect rect, {TextDirection textDirection}) {
-    return _getPath(rect, math.max(0.0, rect.shortestSide / 2.0 - border.width));
+  Path getInnerPath(Rect rect, {TextDirection? textDirection}) {
+    return _getPath(
+        rect, math.max(0.0, rect.shortestSide / 2.0 - border.width));
   }
 
   @override
-  Path getOuterPath(Rect rect, {TextDirection textDirection}) {
+  Path getOuterPath(Rect rect, {TextDirection? textDirection}) {
     return _getPath(rect, math.max(0.0, rect.shortestSide / 2.0));
   }
 
   @override
   ShapeBorder scale(double t) {
-    return PolygonBorder(sides: sides,
-        rotate : rotate,
+    return PolygonBorder(
+        sides: sides,
+        rotate: rotate,
         borderRadius: borderRadius * t,
         border: border.scale(t));
   }
 
   @override
   int get hashCode {
-    return sides.hashCode ^ rotate.hashCode ^ borderRadius.hashCode ^ border.hashCode;
+    return sides.hashCode ^
+        rotate.hashCode ^
+        borderRadius.hashCode ^
+        border.hashCode;
   }
 
   @override
@@ -109,8 +111,11 @@ class PolygonBorder extends ShapeBorder {
       return false;
     }
 
-    final PolygonBorder typedOther = other;
-    return sides == typedOther.sides && rotate == typedOther.rotate &&
-        borderRadius == typedOther.borderRadius && border == typedOther.border;
+    // ignore: test_types_in_equals
+    final PolygonBorder typedOther = other as PolygonBorder;
+    return sides == typedOther.sides &&
+        rotate == typedOther.rotate &&
+        borderRadius == typedOther.borderRadius &&
+        border == typedOther.border;
   }
 }
